Application/requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
scikit-image==0.16.2
tensorflow>=1.3.0
keras>=2.0.8
opencv-python
h5py
imgaug
IPython[all]
xlrd
torch
torchvision
pandas
flask
mysql-connector
Pillow
PyYAML>=5.3.1
tqdm
seaborn
